added async requests

+++ express/controllers/admin.js
@@ -2,9 +2,7 @@ const mongoose = require('mongoose');
 
 const fileHelper = require('../util/file');
 
-const {
-  validationResult
-} = require('express-validator/check');
+const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
 
@@ -59,6 +57,7 @@ exports.postAddProduct = (req, res, next) => {
   }
 
   const imageUrl = image.path;
+
   const product = new Product({
     // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
     title: title,
@@ -175,13 +174,11 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({
-      userId: req.user._id
-    })
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
-      console.log('here');
+      console.log(products);
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
@@ -195,26 +192,23 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
   Product.findById(prodId)
     .then(product => {
       if (!product) {
         return next(new Error('Product not found.'));
       }
       fileHelper.deleteFile(product.imageUrl);
-      return Product.deleteOne({
-        _id: prodId,
-        userId: req.user._id
-      });
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
     })
     .then(() => {
       console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
+      res.status(200).json({
+        message: 'Success',
+      });
     })
     .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
+     res.status(500).json({message: "Deleting Product failed"});
     });
 };
